stream/batch/thumbnail-analyzer/requirements.txt
fastapi
uvicorn
ffmpeg-python
pillow
numpy
